Created the ingredientComponent and moved multiple variables from ItemComponent to IngredientComponent
