+++ source/core/src/main/com/csse3200/game/components/items/ItemComponent.java
@@ -8,21 +8,12 @@ public class ItemComponent extends Component {
     private final String itemName;
     private final int weight;
     private final ItemType itemType;
-    private final int cookTime;
-    private ItemState itemState;
-    private enum ItemState{
-        RAW,
-        COOKED,
-        BURNED
-    }
 
-    public ItemComponent(String itemName, ItemType itemType, int weight, int cookTime) {
+    public ItemComponent(String itemName, ItemType itemType, int weight) {
         this.itemId = UUID.randomUUID().toString();
         this.itemName = itemName;
         this.weight = weight;
         this.itemType = itemType;
-        this.cookTime = cookTime;
-        this.itemState = ItemState.RAW;
     }
 
     public String getItemId() {
@@ -41,15 +32,4 @@ public class ItemComponent extends Component {
         return this.itemType;
     }
 
-    public String getItemState() {
-        return this.itemState.name();
-    }
-
-    public void cookItem() {
-        this.itemState = ItemState.COOKED;
-    }
-
-    public void burnItem() {
-        this.itemState = ItemState.BURNED;
-    }
 }
